+++ Samples~/Collaboration Samples/Scripts/Menus/DirectionalLightMenu.cs
@@ -21,24 +21,15 @@ namespace CavrnusSdk.CollaborationExamples
         [SerializeField] private UISliderWrapper rotationSlider;
         [SerializeField] private UISliderWrapper shadowSlider;
 
-        private CavrnusSpaceConnection spaceConn;
         private List<IDisposable> disposables = new List<IDisposable>();
         private static readonly int Rotation = Shader.PropertyToID("_Rotation");
         
-        private CavrnusLivePropertyUpdate<float> liveSunRotationUpdate;
-        private CavrnusLivePropertyUpdate<float> liveShadowUpdate;
-
         private void Start()
         {
             CavrnusFunctionLibrary.AwaitAnySpaceConnection(spaceConn => {
-                this.spaceConn = spaceConn;
+                rotationSlider.Setup(spaceConn, containerName, rotationPropertyName, new Vector2(0f, 360f));
+                shadowSlider.Setup(spaceConn, containerName, shadowPropertyName, new Vector2(0f, 1f));
 
-                rotationSlider.Slider.minValue = 0f;
-                rotationSlider.Slider.maxValue = 360f;
-                
-                shadowSlider.Slider.minValue = 0f;
-                shadowSlider.Slider.maxValue = 1f;
-                
                 // Sunlight Rotation
                 spaceConn.DefineFloatPropertyDefaultValue(containerName, rotationPropertyName, RenderSettings.skybox.GetFloat(Rotation));
                 disposables.Add(spaceConn.BindFloatPropertyValue(containerName, rotationPropertyName, rotation => {
@@ -47,66 +38,19 @@ namespace CavrnusSdk.CollaborationExamples
                     rotationSlider.Slider.SetValueWithoutNotify(rotation);
                 }));
                 
-                rotationSlider.OnValueUpdated += RotationValueChanged;
-                rotationSlider.OnBeginDragging += RotationDragBegin;
-                rotationSlider.OnEndDragging += RotationDragEnd;
-                
                 // Shadow Strength
                 spaceConn.DefineFloatPropertyDefaultValue(containerName, shadowPropertyName, targetLight.shadowStrength);
                 disposables.Add(spaceConn.BindFloatPropertyValue(containerName, shadowPropertyName, strength => {
                     targetLight.shadowStrength = strength;
                     shadowSlider.Slider.SetValueWithoutNotify(strength);
                 }));
-                shadowSlider.OnValueUpdated += ShadowValueChanged;
-                shadowSlider.OnBeginDragging += ShadowDragBegin;
-                shadowSlider.OnEndDragging += ShadowDragEnd;
             });
-        }
-
-        private void ShadowValueChanged(float val)
-        {
-            liveShadowUpdate?.UpdateWithNewData(val);
-        }
-
-        private void ShadowDragBegin(float val)
-        {
-            liveShadowUpdate ??= spaceConn.BeginTransientFloatPropertyUpdate(containerName, shadowPropertyName, val);
-        }
-
-        private void ShadowDragEnd(float val)
-        {
-            liveShadowUpdate?.Finish();
-            liveShadowUpdate = null;
-        }
-
-        private void RotationDragBegin(float val)
-        {
-            liveSunRotationUpdate ??= spaceConn.BeginTransientFloatPropertyUpdate(containerName, rotationPropertyName, val);
-        }
-
-        private void RotationDragEnd(float val)
-        {
-            liveSunRotationUpdate?.Finish();
-            liveSunRotationUpdate = null;
-        }
-
-        private void RotationValueChanged(float val)
-        {
-            liveSunRotationUpdate?.UpdateWithNewData(val);
         }
 
         private void OnDestroy()
         {
             foreach (var disposable in disposables)
                 disposable.Dispose();
-            
-            rotationSlider.OnValueUpdated -= RotationValueChanged;
-            rotationSlider.OnBeginDragging -= RotationDragBegin;
-            rotationSlider.OnEndDragging -= RotationDragEnd;
-            
-            shadowSlider.OnValueUpdated -= ShadowValueChanged;
-            shadowSlider.OnBeginDragging -= ShadowDragBegin;
-            shadowSlider.OnEndDragging -= ShadowDragEnd;
         }
     }
 }